requirements.txt
PyPDF2
git+https://github.com/huggingface/transformers.git 
sentencepiece
presidio_analyzer
presidio_anonymizer
python-pptx